Move VSCMessages and PyDevdMessages under the helpers package.
