+++ tests/helpers/pydevd/__init__.py
@@ -1,2 +1,3 @@
 from ._fake import FakePyDevd  # noqa
+from ._messages import PyDevdMessages  # noqa
 from ._pydevd import RawMessage  # noqa
