+++ tests/helpers/vsc/__init__.py
@@ -1,2 +1,3 @@
 from ._fake import FakeVSC  # noqa
+from ._messages import VSCMessages  # noqa
 from ._vsc import parse_message, RawMessage, Request, Response, Event  # noqa
